Adapt to moved atomic change project
