+++ bundles/tools.vitruv.extensions.changevisualization/src/tools/vitruv/extensions/changevisualization/tree/decoder/TreeChangeDataSetDecoder.java
@@ -16,7 +16,7 @@ import tools.vitruv.extensions.changevisualization.tree.decoder.feature.Multiple
 import tools.vitruv.extensions.changevisualization.tree.decoder.feature.OldValueNewValueProcessor;
 import tools.vitruv.framework.change.description.PropagatedChange;
 import tools.vitruv.framework.change.description.VitruviusChange;
-import tools.vitruv.framework.change.echange.EChange;
+import tools.vitruv.change.atomic.EChange;
 import tools.vitruv.framework.domains.VitruvDomain;
 
 /**
