+++ bundles/tools.vitruv.extensions.changevisualization/src/tools/vitruv/extensions/changevisualization/tree/decoder/echange/ReplaceChangeDecoder.java
@@ -2,7 +2,7 @@ package tools.vitruv.extensions.changevisualization.tree.decoder.echange;
 
 import java.util.Map;
 
-import tools.vitruv.framework.change.echange.EChange;
+import tools.vitruv.change.atomic.EChange;
 
 /**
  * Decoder for changes with a replace semantic
