+++ bundles/tools.vitruv.extensions.changevisualization/src/tools/vitruv/extensions/changevisualization/tree/decoder/feature/OldValueNewValueProcessor.java
@@ -7,7 +7,7 @@ import java.util.Vector;
 import javax.swing.tree.DefaultMutableTreeNode;
 
 import tools.vitruv.extensions.changevisualization.tree.FeatureNode;
-import tools.vitruv.framework.change.echange.EChange;
+import tools.vitruv.change.atomic.EChange;
 
 /**
  * This multiple feature processors replaces pairs of oldValue and newValue structural features.
