+++ features/tools.vitruv.extensions.changevisualization.feature/feature.xml
@@ -26,7 +26,7 @@
       <import plugin="org.eclipse.emf.common"/>
       <import plugin="org.eclipse.emf.ecore"/>
       <import plugin="tools.vitruv.framework.change"/>
-      <import plugin="tools.vitruv.framework.change.echange"/>
+      <import plugin="tools.vitruv.change.atomic"/>
       <import plugin="tools.vitruv.framework.domains"/>
       <import plugin="tools.vitruv.framework.vsum"/>
    </requires>
