+++ tests/tools.vitruv.domains.emf.ui.tests/src/tools/vitruv/domains/emf/monitorededitor/monitor/SynchronizingMonitoredEmfEditorImplTests.java
@@ -40,7 +40,7 @@ import tools.vitruv.domains.emf.monitorededitor.test.utils.EnsureNotExecuted;
 import tools.vitruv.domains.emf.monitorededitor.tools.EclipseAdapterProvider;
 import tools.vitruv.domains.emf.monitorededitor.tools.IEclipseAdapter;
 import tools.vitruv.framework.change.description.VitruviusChange;
-import tools.vitruv.framework.change.echange.feature.FeatureEChange;
+import tools.vitruv.change.atomic.feature.FeatureEChange;
 
 public class SynchronizingMonitoredEmfEditorImplTests extends BasicTestCase {
     private EclipseMock eclipseCtrl;
