+++ tests/tools.vitruv.domains.emf.ui.tests/src/tools/vitruv/domains/emf/monitorededitor/monitor/VitruviusEMFEditorMonitorImplTests.java
@@ -39,7 +39,7 @@ import tools.vitruv.domains.emf.monitorededitor.tools.EclipseAdapterProvider;
 import tools.vitruv.domains.emf.monitorededitor.tools.IEclipseAdapter;
 import tools.vitruv.framework.change.description.CompositeContainerChange;
 import tools.vitruv.framework.change.description.VitruviusChange;
-import tools.vitruv.framework.change.echange.feature.FeatureEChange;
+import tools.vitruv.change.atomic.feature.FeatureEChange;
 
 public class VitruviusEMFEditorMonitorImplTests extends BasicTestCase {
     private EclipseMock eclipseMockCtrl;
