+++ tests/tools.vitruv.domains.emf.ui.tests/src/tools/vitruv/domains/emf/monitorededitor/test/utils/ChangeAssert.java
@@ -15,9 +15,9 @@ import java.util.Collection;
 
 import org.eclipse.emf.ecore.EStructuralFeature;
 
-import tools.vitruv.framework.change.echange.EChange;
-import tools.vitruv.framework.change.echange.feature.FeatureEChange;
-import tools.vitruv.framework.change.echange.feature.attribute.ReplaceSingleValuedEAttribute;
+import tools.vitruv.change.atomic.EChange;
+import tools.vitruv.change.atomic.feature.FeatureEChange;
+import tools.vitruv.change.atomic.feature.attribute.ReplaceSingleValuedEAttribute;
 import tools.vitruv.framework.change.description.VitruviusChange;
 
 public final class ChangeAssert {
