+++ tests/tools.vitruv.framework.views.tests/src/tools/vitruv/framework/views/impl/ChangeDerivingViewTest.java
@@ -28,11 +28,11 @@ import org.mockito.MockitoAnnotations;
 import allElementTypes.NonRoot;
 import allElementTypes.Root;
 import tools.vitruv.framework.change.description.VitruviusChange;
-import tools.vitruv.framework.change.echange.eobject.EobjectPackage;
-import tools.vitruv.framework.change.echange.feature.attribute.ReplaceSingleValuedEAttribute;
-import tools.vitruv.framework.change.echange.root.InsertRootEObject;
-import tools.vitruv.framework.change.echange.root.RootFactory;
-import tools.vitruv.framework.change.echange.root.RootPackage;
+import tools.vitruv.change.atomic.eobject.EobjectPackage;
+import tools.vitruv.change.atomic.feature.attribute.ReplaceSingleValuedEAttribute;
+import tools.vitruv.change.atomic.root.InsertRootEObject;
+import tools.vitruv.change.atomic.root.RootFactory;
+import tools.vitruv.change.atomic.root.RootPackage;
 import tools.vitruv.framework.views.ChangeableViewSource;
 import tools.vitruv.framework.views.ModifiableViewSelection;
 import tools.vitruv.framework.views.changederivation.DefaultStateBasedChangeResolutionStrategy;
