+++ tests/tools.vitruv.framework.views.tests/src/tools/vitruv/framework/views/impl/ChangeRecordingViewTest.java
@@ -31,18 +31,18 @@ import allElementTypes.AllElementTypesPackage;
 import allElementTypes.NonRoot;
 import allElementTypes.Root;
 import tools.vitruv.framework.change.description.VitruviusChange;
-import tools.vitruv.framework.change.echange.EChange;
-import tools.vitruv.framework.change.echange.eobject.CreateEObject;
-import tools.vitruv.framework.change.echange.eobject.DeleteEObject;
-import tools.vitruv.framework.change.echange.eobject.EobjectPackage;
-import tools.vitruv.framework.change.echange.feature.FeaturePackage;
-import tools.vitruv.framework.change.echange.feature.attribute.AttributeFactory;
-import tools.vitruv.framework.change.echange.feature.attribute.AttributePackage;
-import tools.vitruv.framework.change.echange.feature.attribute.ReplaceSingleValuedEAttribute;
-import tools.vitruv.framework.change.echange.feature.reference.ReplaceSingleValuedEReference;
-import tools.vitruv.framework.change.echange.root.InsertRootEObject;
-import tools.vitruv.framework.change.echange.root.RootFactory;
-import tools.vitruv.framework.change.echange.root.RootPackage;
+import tools.vitruv.change.atomic.EChange;
+import tools.vitruv.change.atomic.eobject.CreateEObject;
+import tools.vitruv.change.atomic.eobject.DeleteEObject;
+import tools.vitruv.change.atomic.eobject.EobjectPackage;
+import tools.vitruv.change.atomic.feature.FeaturePackage;
+import tools.vitruv.change.atomic.feature.attribute.AttributeFactory;
+import tools.vitruv.change.atomic.feature.attribute.AttributePackage;
+import tools.vitruv.change.atomic.feature.attribute.ReplaceSingleValuedEAttribute;
+import tools.vitruv.change.atomic.feature.reference.ReplaceSingleValuedEReference;
+import tools.vitruv.change.atomic.root.InsertRootEObject;
+import tools.vitruv.change.atomic.root.RootFactory;
+import tools.vitruv.change.atomic.root.RootPackage;
 import tools.vitruv.framework.views.ChangeableViewSource;
 import tools.vitruv.framework.views.ModifiableViewSelection;
 import tools.vitruv.testutils.RegisterMetamodelsInStandalone;
